requirements.txt
ai2thor==2.1.0
-e git+https://github.com/alfworld/alfworld.git@2f8c6608986bbf62b2020fb2b4489449a2df3dd4#egg=alfworld
annotated-types==0.6.0
anyio==4.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
distro==1.9.0
exceptiongroup==1.2.0
fast-downward-textworld==20.6.1
filelock==3.13.1
Flask==2.1.3
fsspec==2024.2.0
h11==0.14.0
hashids==1.3.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
jericho==3.1.2
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.5
mementos==1.3.1
more-itertools==10.2.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.3
openai==1.13.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
preshed==3.0.9
progressbar2==4.3.2
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
TatSu==5.8.3
termcolor==2.4.0
textworld==1.6.1
thinc==8.2.2
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.38.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==2.0.3
zipp==3.17.0
